- Cambio de imagen a  Text en código SwiftUI - Agregue  nueva propiedad "cantidad" en Estructura Comida -Cambie nombre de estructura ListaComidaView a CarritoView -Agreue List en el ForEach - En Preview agregue el .prederredColorScheme(.dark) para modo oscuro
