+++ POS/BaseDeDatosComida.swift
@@ -17,4 +17,6 @@ class BaseDeDatosComida{
         comidas.append(comida)
     }
     
+    
+    
 }
